Added get request to product service
